# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts Manager App",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "Contacts operations"
  },
  "servers": [
    {
      "url": "https://hw7-swagger-qzew.onrender.com"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
  "tags": [
    {
      "name": "Contacts",
      "description": "Операції з контактами"
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Отримати всі контакти",
        "operationId": "getAllContacts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "name"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "example": "asc"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Фільтрація за типом контакту",
            "schema": {
              "type": "string",
              "example": "work"
            }
          },
          {
            "name": "isFavourite",
            "in": "query",
            "description": "Фільтрація за позначкою улюбленості (true/false)",
            "schema": {
              "type": "string",
              "example": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список контактів",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Contacts retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Contact"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Створити новий контакт",
        "operationId": "createContact",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "phoneNumber"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Ім'я контакту",
                    "example": "Іван Петров"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Номер телефону",
                    "example": "+380975554532"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email",
                    "example": "ivanpele@gmail.com"
                  },
                  "contactType": {
                    "type": "string",
                    "enum": ["work", "personal", "home"],
                    "description": "Тип контакту",
                    "example": "work"
                  },
                  "photo": {
                    "type": "string",
                    "description": "URL або base64 зображення",
                    "example": ""
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Позначка, чи є контакт улюбленим",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Успішно створено контакт",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessCreatedResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Отримати контакт за ID",
        "operationId": "getContactById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67ca67e7ae7f10c25b566655"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Знайдено контакт",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "tags": ["Contacts"],
        "summary": "Оновити контакт",
        "operationId": "updateContact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67ca67e7ae7f10c25b566655"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Ім'я контакту"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Номер телефону"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email"
                  },
                  "contactType": {
                    "type": "string",
                    "enum": ["work", "personal", "home"]
                  },
                  "photo": {
                    "type": "string",
                    "description": "URL або base64 зображення"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Позначка, чи є контакт улюбленим"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успішно оновлено контакт",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Видалити контакт",
        "operationId": "deleteContact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Контакт успішно видалено"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Contact": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "67ca67e7ae7f10c25b566655"
          },
          "name": {
            "type": "string",
            "example": "Іван Петров"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+380975554532"
          },
          "email": {
            "type": "string",
            "example": "ivanpele@gmail.com"
          },
          "contactType": {
            "type": "string",
            "enum": ["work", "personal", "home"]
          },
          "photo": {
            "type": "string",
            "example": ""
          },
          "isFavourite": {
            "type": "boolean",
            "example": true
          },
          "userId": {
            "type": "string",
            "example": "67bfa7a9925aebdb493b5b33"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "type": "object",
            "$ref": "#/components/schemas/Contact"
          }
        }
      },
      "SuccessCreatedResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 201
          },
          "message": {
            "type": "string",
            "example": "Successfully created a contact!"
          },
          "data": {
            "type": "object",
            "$ref": "#/components/schemas/Contact"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Невірний запит",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Bad request"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Неавторизовано",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Не знайдено",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Not found"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "Внутрішня помилка сервера",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    }
  }
}